refactor: create infrastructure module and move EventSourcing Group file there
