+++ build.gradle.kts
@@ -28,10 +28,6 @@ allprojects {
             implementation(scala.library)
             implementation(cats.core)
             implementation(akka.actor.typed)
-            implementation(akka.cluster.typed)
-            implementation(akka.cluster.sharding.typed)
-            implementation(akka.persistence.cassandra)
-            implementation(akka.projection.core)
             implementation(logback.classic)
 
             testImplementation(bundles.scala.testing)
