+++ domain/build.gradle.kts
@@ -1,7 +1,3 @@
 dependencies {
     api(project(":common"))
-    with(rootProject.libs) {
-        implementation(akka.serialization.jackson)
-        implementation(akka.persistence.typed)
-    }
 }
