+++ infrastructure/build.gradle.kts
@@ -1,0 +1,11 @@
+dependencies {
+    api(project(":common"))
+    with(rootProject.libs) {
+        implementation(akka.cluster.typed)
+        implementation(akka.cluster.sharding.typed)
+        implementation(akka.persistence.cassandra)
+        implementation(akka.projection.core)
+        implementation(akka.serialization.jackson)
+        implementation(akka.persistence.typed)
+    }
+}
